Update admin page for members and refactor api routes
